Switch to doubles

+++ Foxtaur/Foxtaur.Desktop/Controls/Renderer/Implementations/Generators/MapSegmentGenerator.cs
@@ -91,10 +91,10 @@ public class MapSegmentGenerator : IMapSegmentGenerator
     /// Returns pair of generated points indices 
     /// </summary>
     private ValueTuple<uint, uint> GenerateAndAddPointsPair(Mesh mesh,
-        float p0Lat,
-        float p0Lon,
-        float p1Lat,
-        float p1Lon,
+        double p0Lat,
+        double p0Lon,
+        double p1Lat,
+        double p1Lon,
         ZoomLevel desiredZoomLevel,
         HighResMapFragment fragment)
     {
